TEMP: display "settings loader triple" at startup for debugging.
This shows the size of three different Collections that are created by neo4j and used when configuring a database.  These are created using "service loaders", something I never learned about before and wish I never had to.  When run with mvn exec:java, the sizes are (15, 2, 4), but with the JAR it's (1, 1, 1).

I think that loading these in correctly should fix the problem and allow a 2.0 release.