Completely remove neo4j.
neo4j is a graph database management system, which was previously used for
handling phrases used and their usage counts.  Graph databases made sense for
this, since they can be used to track things like what phrases are commonly used
together or after each other, and was great for a dissertation project, but in
practice this wasn't developed very far, and the "phrases used together" feature
doesn't currently justify the complexity of having a whole DBMS that must be
imported and maintained.

neo4j has been a real pain lately.  The old version only supports Java 11, and
the new version only 17.  And there's some weird bug that occurs occasionally
where the database is corrupted, making it impossible to load an assignment.

This commit replaces neo4j with a simple system that stores phrase details in
json and writes them to a single file.  There's no tracking for "phrases used
together", so this feature will stop working and should now be removed from the
GUI.

One day this should perhaps be reverted and the lost features readded.  For now,
this will be easier.